Get rid of ExoPlayerFactory
